Added permissions for all role commands

+++ commands/delrole.js
@@ -1,3 +1,5 @@
+const { GuildMember } = require("discord.js");
+
 module.exports = {
 	name: 'delrole',
     description: 'Removes a role to the server',
@@ -7,12 +9,12 @@ module.exports = {
     guildOnly: true,
 	execute(message, args, client) {
         let roleFind = message.guild.roles.cache.find(r => r.name === args[0]);
-        if (roleFind) {
+        if (roleFind && message.member.hasPermission('ADMINISTRATOR')) {
             roleFind.delete();
             message.channel.send(`The role ${args[0]} has been deleted`)
         }
         else {
-            message.channel.send('That role does not exist')
+            message.channel.send('That role does not exist or you do not have permission to preform that command')
         }
 	},
 };``
